+++ src/main/java/org/apache/skywalking/banyandb/v1/client/MeasureBulkWriteProcessor.java
@@ -22,9 +22,9 @@ import io.grpc.stub.StreamObserver;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.skywalking.banyandb.measure.v1.BanyandbMeasure;
 import org.apache.skywalking.banyandb.measure.v1.MeasureServiceGrpc;
-import org.apache.skywalking.banyandb.v1.client.grpc.GRPCStreamServiceStatus;
 
 import javax.annotation.concurrent.ThreadSafe;
+import java.util.concurrent.CompletableFuture;
 
 /**
  * MeasureBulkWriteProcessor works for measure flush.
@@ -52,7 +52,7 @@ public class MeasureBulkWriteProcessor extends AbstractBulkWriteProcessor<Banyan
 
     @Override
     protected StreamObserver<BanyandbMeasure.WriteRequest> buildStreamObserver(MeasureServiceGrpc.MeasureServiceStub stub,
-                                                                               GRPCStreamServiceStatus status) {
+                                                                               CompletableFuture<Void> batch) {
         return stub.write(new StreamObserver<BanyandbMeasure.WriteResponse>() {
             @Override
             public void onNext(BanyandbMeasure.WriteResponse writeResponse) {
@@ -61,13 +61,13 @@ public class MeasureBulkWriteProcessor extends AbstractBulkWriteProcessor<Banyan
 
             @Override
             public void onError(Throwable t) {
-                status.finished();
+                batch.completeExceptionally(t);
                 log.error("Error occurs in flushing measures", t);
             }
 
             @Override
             public void onCompleted() {
-                status.finished();
+                batch.complete(null);
             }
         });
     }