+++ src/main/java/org/apache/skywalking/banyandb/v1/client/grpc/MetadataClient.java
@@ -21,6 +21,7 @@ package org.apache.skywalking.banyandb.v1.client.grpc;
 import com.google.protobuf.GeneratedMessageV3;
 import io.grpc.stub.AbstractBlockingStub;
 import org.apache.skywalking.banyandb.v1.client.grpc.exception.BanyanDBException;
+import org.apache.skywalking.banyandb.v1.client.metadata.ResourceExist;
 import org.apache.skywalking.banyandb.v1.client.metadata.NamedSchema;
 
 import java.util.List;
@@ -73,6 +74,16 @@ public abstract class MetadataClient<STUB extends AbstractBlockingStub<STUB>, P 
      * @throws BanyanDBException a wrapped exception to the underlying gRPC calls
      */
     public abstract S get(String group, String name) throws BanyanDBException;
+
+    /**
+     * Check whether a schema exists
+     *
+     * @param group the group of the schema to be found
+     * @param name the name of the schema to be found
+     * @return whether resource exists
+     * @throws BanyanDBException a wrapped exception to the underlying gRPC calls
+     */
+    public  abstract ResourceExist exist(String group, String name) throws BanyanDBException;
 
     /**
      * List all schemas with the same group name