+++ src/test/java/org/apache/skywalking/banyandb/v1/client/ITBanyanDBPropertyTests.java
@@ -21,8 +21,8 @@ package org.apache.skywalking.banyandb.v1.client;
 import io.grpc.Status;
 import org.apache.skywalking.banyandb.v1.client.grpc.exception.BanyanDBException;
 import org.apache.skywalking.banyandb.v1.client.metadata.Catalog;
-import org.apache.skywalking.banyandb.v1.client.metadata.Duration;
 import org.apache.skywalking.banyandb.v1.client.metadata.Group;
+import org.apache.skywalking.banyandb.v1.client.metadata.IntervalRule;
 import org.apache.skywalking.banyandb.v1.client.metadata.Property;
 import org.junit.After;
 import org.junit.Assert;
@@ -39,7 +39,9 @@ public class ITBanyanDBPropertyTests extends BanyanDBClientTestCI {
     public void setUp() throws IOException, BanyanDBException, InterruptedException {
         super.setUpConnection();
         Group expectedGroup = this.client.define(
-                Group.create("default", Catalog.STREAM, 2, 0, Duration.ofDays(7))
+                Group.create("default", Catalog.STREAM, 2, IntervalRule.create(IntervalRule.Unit.HOUR, 4),
+                        IntervalRule.create(IntervalRule.Unit.DAY, 1),
+                        IntervalRule.create(IntervalRule.Unit.DAY, 7))
         );
         Assert.assertNotNull(expectedGroup);
     }
@@ -54,7 +56,7 @@ public class ITBanyanDBPropertyTests extends BanyanDBClientTestCI {
         Property property = Property.create("default", "sw", "ui_template")
                 .addTag(TagAndValue.newStringTag("name", "hello"))
                 .build();
-        this.client.save(property);
+        Assert.assertTrue(this.client.apply(property).created());
 
         await().atMost(10, TimeUnit.SECONDS).untilAsserted(() -> {
             Property gotProperty = client.findProperty("default", "sw", "ui_template");
@@ -68,7 +70,7 @@ public class ITBanyanDBPropertyTests extends BanyanDBClientTestCI {
         Property property = Property.create("default", "sw", "ui_template")
                 .addTag(TagAndValue.newStringTag("name", "hello"))
                 .build();
-        this.client.save(property);
+        Assert.assertTrue(this.client.apply(property).created());
 
         await().atMost(10, TimeUnit.SECONDS).untilAsserted(() -> {
             Property gotProperty = client.findProperty("default", "sw", "ui_template");
@@ -76,7 +78,7 @@ public class ITBanyanDBPropertyTests extends BanyanDBClientTestCI {
             Assert.assertEquals(property, gotProperty);
         });
 
-        Assert.assertTrue(this.client.deleteProperty("default", "sw", "ui_template"));
+        Assert.assertTrue(this.client.deleteProperty("default", "sw", "ui_template").deleted());
 
         try {
             client.findProperty("default", "sw", "ui_template");
@@ -91,12 +93,12 @@ public class ITBanyanDBPropertyTests extends BanyanDBClientTestCI {
         Property property1 = Property.create("default", "sw", "ui_template")
                 .addTag(TagAndValue.newStringTag("name", "hello"))
                 .build();
-        this.client.save(property1);
+        Assert.assertTrue(this.client.apply(property1).created());
 
         Property property2 = Property.create("default", "sw", "ui_template")
                 .addTag(TagAndValue.newStringTag("name", "world"))
                 .build();
-        this.client.save(property2);
+        Assert.assertFalse(this.client.apply(property2).created());
 
         await().atMost(10, TimeUnit.SECONDS).untilAsserted(() -> {
             Property gotProperty = client.findProperty("default", "sw", "ui_template");