+++ CHANGES.md
@@ -8,7 +8,15 @@ Release Notes.
 
 ### Features
 
+- Support `indexed_only` flag to the tag specification.
+- Support `Analyzer` to the index rule.
+- Add `exist` endpoints to the metadata registry.
+- Set `CompletableFuture<Void>` to the return type of write processor.
+- Refactor property operations.
+
 ### Bugs
+
+- Fix UTs failures with JDK16,17(https://github.com/apache/skywalking/issues/9771)
 
 
 0.1.0